feat: moving events out of shared interfaces
