+++ contracts/bond/BondPortal.sol
@@ -19,8 +19,6 @@ import "../Version.sol";
  * @dev Orchestrates the various Bond contracts to provide a single function to aggregate the various calls.
  */
 interface BondPortal {
-    event CreateDao(uint256 indexed id, address indexed treasury);
-
     /**
      * @notice Initialises a new DAO with essential configuration.
      *
